+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.279] - 2024-02-25 - csv export bugfix
+
+## [0.0.278] - 2024-02-25 - csv export bugfix
+
 ## [0.0.277] - 2024-02-25 - bugfix scriptblock invocation
 
 ## [0.0.276] - 2024-02-25 - regex bugfix